RaspberryPi/requirements.txt
PyYAML==3.12
awscli==1.11.120
botocore==1.5.83
chardet==2.3.0
colorama==0.3.2
docutils==0.13.1
html5lib==0.999
jmespath==0.9.3
pyasn1==0.2.3
pyserial==3.3
python-dateutil==2.6.1
requests==2.4.3
rsa==3.4.2
s3transfer==0.1.10
six==1.8.0
urllib3==1.9.1
wheel==0.24.0